refactor: Move mock repository
